Refactored People page, added "imageUrl" shorthand
