+++ src/common/imageUrl.js
@@ -1,0 +1,1 @@
+export const imageUrl = "https://image.tmdb.org/t/p/original";
